ch(application routes and controllers):  implement LFA server feedback
- create controllers for routes to make entry point file simpler and reflect the MVC pattern
- restructure server folder as instructed by LFA

[Delivers #159298058]
