--- .nyc_output/81c73224f3c33bc6f776cfa4c225d6f0.json
@@ -1,1 +1,0 @@
-{"C:\\Users\\TOSHIBA\\Desktop\\Andela-bootcamp\\MyDiary\\server\\src\\app.js":{"path":"C:\\Users\\TOSHIBA\\Desktop\\Andela-bootcamp\\MyDiary\\server\\src\\app.js","statementMap":{"0":{"start":{"line":8,"column":12},"end":{"line":8,"column":21}},"1":{"start":{"line":11,"column":0},"end":{"line":11,"column":23}},"2":{"start":{"line":14,"column":0},"end":{"line":14,"column":27}},"3":{"start":{"line":15,"column":0},"end":{"line":15,"column":52}},"4":{"start":{"line":18,"column":0},"end":{"line":21,"column":4}},"5":{"start":{"line":18,"column":27},"end":{"line":21,"column":2}},"6":{"start":{"line":23,"column":0},"end":{"line":25,"column":4}},"7":{"start":{"line":23,"column":29},"end":{"line":25,"column":2}},"8":{"start":{"line":28,"column":0},"end":{"line":34,"column":3}},"9":{"start":{"line":29,"column":2},"end":{"line":33,"column":4}},"10":{"start":{"line":37,"column":0},"end":{"line":56,"column":3}},"11":{"start":{"line":38,"column":17},"end":{"line":38,"column":27}},"12":{"start":{"line":39,"column":2},"end":{"line":55,"column":3}},"13":{"start":{"line":40,"column":23},"end":{"line":42,"column":6}},"14":{"start":{"line":41,"column":6},"end":{"line":41,"column":33}},"15":{"start":{"line":44,"column":4},"end":{"line":52,"column":5}},"16":{"start":{"line":45,"column":6},"end":{"line":49,"column":8}},"17":{"start":{"line":51,"column":6},"end":{"line":51,"column":29}},"18":{"start":{"line":54,"column":4},"end":{"line":54,"column":27}},"19":{"start":{"line":59,"column":0},"end":{"line":80,"column":3}},"20":{"start":{"line":60,"column":2},"end":{"line":79,"column":3}},"21":{"start":{"line":61,"column":4},"end":{"line":66,"column":7}},"22":{"start":{"line":67,"column":4},"end":{"line":76,"column":6}},"23":{"start":{"line":78,"column":4},"end":{"line":78,"column":27}},"24":{"start":{"line":84,"column":0},"end":{"line":109,"column":3}},"25":{"start":{"line":85,"column":2},"end":{"line":108,"column":3}},"26":{"start":{"line":86,"column":19},"end":{"line":86,"column":29}},"27":{"start":{"line":88,"column":23},"end":{"line":90,"column":6}},"28":{"start":{"line":89,"column":6},"end":{"line":89,"column":33}},"29":{"start":{"line":92,"column":4},"end":{"line":105,"column":5}},"30":{"start":{"line":93,"column":6},"end":{"line":102,"column":8}},"31":{"start":{"line":104,"column":6},"end":{"line":104,"column":29}},"32":{"start":{"line":107,"column":4},"end":{"line":107,"column":27}},"33":{"start":{"line":112,"column":13},"end":{"line":112,"column":37}},"34":{"start":{"line":114,"column":0},"end":{"line":116,"column":3}},"35":{"start":{"line":115,"column":2},"end":{"line":115,"column":45}}},"fnMap":{"0":{"name":"(anonymous_0)","decl":{"start":{"line":18,"column":13},"end":{"line":18,"column":14}},"loc":{"start":{"line":18,"column":27},"end":{"line":21,"column":2}},"line":18},"1":{"name":"(anonymous_1)","decl":{"start":{"line":23,"column":15},"end":{"line":23,"column":16}},"loc":{"start":{"line":23,"column":29},"end":{"line":25,"column":2}},"line":23},"2":{"name":"(anonymous_2)","decl":{"start":{"line":28,"column":27},"end":{"line":28,"column":28}},"loc":{"start":{"line":28,"column":41},"end":{"line":34,"column":1}},"line":28},"3":{"name":"(anonymous_3)","decl":{"start":{"line":37,"column":31},"end":{"line":37,"column":32}},"loc":{"start":{"line":37,"column":45},"end":{"line":56,"column":1}},"line":37},"4":{"name":"(anonymous_4)","decl":{"start":{"line":40,"column":36},"end":{"line":40,"column":37}},"loc":{"start":{"line":40,"column":47},"end":{"line":42,"column":5}},"line":40},"5":{"name":"(anonymous_5)","decl":{"start":{"line":59,"column":28},"end":{"line":59,"column":29}},"loc":{"start":{"line":59,"column":42},"end":{"line":80,"column":1}},"line":59},"6":{"name":"(anonymous_6)","decl":{"start":{"line":84,"column":31},"end":{"line":84,"column":32}},"loc":{"start":{"line":84,"column":45},"end":{"line":109,"column":1}},"line":84},"7":{"name":"(anonymous_7)","decl":{"start":{"line":88,"column":36},"end":{"line":88,"column":37}},"loc":{"start":{"line":88,"column":47},"end":{"line":90,"column":5}},"line":88},"8":{"name":"(anonymous_8)","decl":{"start":{"line":114,"column":17},"end":{"line":114,"column":18}},"loc":{"start":{"line":114,"column":23},"end":{"line":116,"column":1}},"line":114}},"branchMap":{"0":{"loc":{"start":{"line":39,"column":2},"end":{"line":55,"column":3}},"type":"if","locations":[{"start":{"line":39,"column":2},"end":{"line":55,"column":3}},{"start":{"line":39,"column":2},"end":{"line":55,"column":3}}],"line":39},"1":{"loc":{"start":{"line":44,"column":4},"end":{"line":52,"column":5}},"type":"if","locations":[{"start":{"line":44,"column":4},"end":{"line":52,"column":5}},{"start":{"line":44,"column":4},"end":{"line":52,"column":5}}],"line":44},"2":{"loc":{"start":{"line":60,"column":2},"end":{"line":79,"column":3}},"type":"if","locations":[{"start":{"line":60,"column":2},"end":{"line":79,"column":3}},{"start":{"line":60,"column":2},"end":{"line":79,"column":3}}],"line":60},"3":{"loc":{"start":{"line":60,"column":6},"end":{"line":60,"column":86}},"type":"binary-expr","locations":[{"start":{"line":60,"column":6},"end":{"line":60,"column":20}},{"start":{"line":60,"column":24},"end":{"line":60,"column":44}},{"start":{"line":60,"column":48},"end":{"line":60,"column":64}},{"start":{"line":60,"column":68},"end":{"line":60,"column":86}}],"line":60},"4":{"loc":{"start":{"line":85,"column":2},"end":{"line":108,"column":3}},"type":"if","locations":[{"start":{"line":85,"column":2},"end":{"line":108,"column":3}},{"start":{"line":85,"column":2},"end":{"line":108,"column":3}}],"line":85},"5":{"loc":{"start":{"line":85,"column":6},"end":{"line":85,"column":86}},"type":"binary-expr","locations":[{"start":{"line":85,"column":6},"end":{"line":85,"column":20}},{"start":{"line":85,"column":24},"end":{"line":85,"column":44}},{"start":{"line":85,"column":48},"end":{"line":85,"column":64}},{"start":{"line":85,"column":68},"end":{"line":85,"column":86}}],"line":85},"6":{"loc":{"start":{"line":92,"column":4},"end":{"line":105,"column":5}},"type":"if","locations":[{"start":{"line":92,"column":4},"end":{"line":105,"column":5}},{"start":{"line":92,"column":4},"end":{"line":105,"column":5}}],"line":92},"7":{"loc":{"start":{"line":112,"column":13},"end":{"line":112,"column":37}},"type":"binary-expr","locations":[{"start":{"line":112,"column":13},"end":{"line":112,"column":29}},{"start":{"line":112,"column":33},"end":{"line":112,"column":37}}],"line":112}},"s":{"0":1,"1":1,"2":1,"3":1,"4":1,"5":0,"6":1,"7":0,"8":1,"9":1,"10":1,"11":3,"12":3,"13":3,"14":8,"15":3,"16":1,"17":2,"18":0,"19":1,"20":2,"21":1,"22":1,"23":1,"24":1,"25":3,"26":2,"27":2,"28":4,"29":2,"30":1,"31":1,"32":1,"33":1,"34":1,"35":1},"f":{"0":0,"1":0,"2":1,"3":3,"4":8,"5":2,"6":3,"7":4,"8":1},"b":{"0":[3,0],"1":[1,2],"2":[1,1],"3":[2,1,1,1],"4":[2,1],"5":[3,2,2,2],"6":[1,1],"7":[1,1]},"_coverageSchema":"332fd63041d2c1bcb487cc26dd0d5f7d97098a6c","hash":"4addcac0255133c723602ffe1edf91c1fd627c02","contentHash":"62b9c17024c793d297b6b3dd851a411d_12.0.2"},"C:\\Users\\TOSHIBA\\Desktop\\Andela-bootcamp\\MyDiary\\server\\data\\dataEntries.js":{"path":"C:\\Users\\TOSHIBA\\Desktop\\Andela-bootcamp\\MyDiary\\server\\data\\dataEntries.js","statementMap":{"0":{"start":{"line":1,"column":16},"end":{"line":15,"column":1}}},"fnMap":{},"branchMap":{},"s":{"0":1},"f":{},"b":{},"_coverageSchema":"332fd63041d2c1bcb487cc26dd0d5f7d97098a6c","hash":"52850abb3a111f4581a8d6368b97a5907c51166b","contentHash":"aafdb5c467b30f3a49e0a53c2a73565f_12.0.2"}}
