--- bin/dev
@@ -1,3 +1,0 @@
-
-require('../dist/src/app');
-
