+++ package.json
@@ -4,12 +4,12 @@
   "description": "MyDiary is an online journal where users can pen down their thoughts and feelings and access it anytime they want.",
   "main": "index.js",
   "scripts": {
-    "start": "node dist/src/app.js",
+    "start": "node dist/routes/src/app.js",
     "test": "cross-env NODE_ENV=test nyc mocha server/**/*.test.js --exit",
     "test-watch": "nodemon --exec \"npm test\"",
     "clean": "rm -rf dist",
     "build": "npm run clean && mkdir dist && babel server -s -d dist",
-    "start:dev": "npm run build && node dist/src/app",
+    "start:dev": "npm run build && node dist/routes/src/app",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "engines": {
