+++ server/controllers/getEntries.js
@@ -1,0 +1,12 @@
+import entries from '../models/data/dataEntries';
+
+
+const getEntries = (req, res) => {
+  res.status(200).json(
+    {
+      entries,
+    },
+  );
+};
+
+export default getEntries;
