--- server/index.js
@@ -1,3 +1,0 @@
-import test from './test';
-
-test('production IS WORKING');
