--- server/test.js
@@ -1,5 +1,0 @@
-const test = (msg) => {
-  console.log(msg);
-}
-
-export default test;
