+++ server/tests/routes/entries/getEntries.test.js
@@ -1,7 +1,8 @@
 import request from 'supertest';
 import { expect } from 'chai';
 
-import app from '../../../src/app';
+
+import app from '../../../routes/src/app';
 
 
 describe('/GET all entries', () => {
