+++ server/tests/routes/entries/getEntry.test.js
@@ -1,8 +1,9 @@
 import request from 'supertest';
 import { expect } from 'chai';
 
-import app from '../../../src/app';
-import entries from '../../../data/dataEntries';
+
+import app from '../../../routes/src/app';
+import entries from '../../../models/data/dataEntries';
 
 
 describe('/GET an entry', () => {
